CFG-869: Config Inspect Must change the AppName and AppSecret
This commit supports both inspector and manager to be used in
different environments. The config now sets up the proper
values after identifying the environment it is running on.

Since the local environment doesn't have browserauth->private,
we can use the config manager token for now.

This allows calling the Config Proxy with BrowserAuth when running
in the deployed environments talking to the Manager and talking to
the Config Manager locally until we have a deployed version.