CFG-876: Making single transactionID + Refactoring fetch with CORS
All the requests are now with a transaction Id created during the
bootstrap. The requests to Github don't take intuit_tid.

CORS is enabled when the app is deployed by other "portal". That
way, any cookies set in the browser will be also submitted to the
URLs from fetch.

This is the pre-work for integration.

*	modified:   app/components/App.jsx
- Adding the transaction ID to be created when mounting
  or rendering.
- Proxying the transactionId to all components

*	modified:   app/components/UserInputs.jsx
- Updating the fetch requests with a specific request object
  with all the parameters needed.

*	modified:   app/components/Views.jsx
- Updating the views fetch requests with a specific object
  with all the parameters for the config service.

*	modified:   app/conf.js
- Centralizing the transactionId method